micro: extract agent model picker
