+++ frontend/src/app/agent/_components/agent-workspace-shell.tsx
@@ -1,17 +1,15 @@
 "use client";
 
-import { useMemo, useRef, useState, useSyncExternalStore, type ReactNode } from "react";
-import { loadSavedControllers } from "@/lib/controllers";
-import { getStoredBackendUrl } from "@/lib/backend-url";
+import type { ReactNode } from "react";
 import { useSearchParams } from "next/navigation";
 import { triggerAddProjectFlow } from "@/ui/projects-nav-section";
-import { ChevronDownIcon, CloseIcon, PlusIcon } from "@/ui/icons";
+import { AgentModelPicker } from "@/ui/agent-model-picker";
+import { CloseIcon, PlusIcon } from "@/ui/icons";
 import type { WorkspaceDispatch } from "@/lib/agent/workspace/effects";
 import type { AgentModel, PaneId, PaneState, WorkspaceState } from "@/lib/agent/workspace/types";
 import { useProjects, type ProjectsContextValue } from "@/lib/agent/projects/context";
 import { useTools } from "@/lib/agent/tools/context";
 import type { Project } from "@/lib/agent/projects/types";
-import { useClickOutside } from "@/hooks/use-click-outside";
 import { useAgentWorkspaceNavigationEffects } from "@/hooks/agent/use-agent-workspace-navigation-effects";
 import { useActiveCanvasSessionEffects } from "@/hooks/agent/use-active-canvas-session-effects";
 import { activeSession, focusedSession } from "@/lib/agent/sessions/selectors";
@@ -292,7 +290,7 @@ function renderWorkspacePane({
       gitSummary={view.gitSummary}
       onInitGit={handles.initGitForActiveProject}
       modelSelector={
-        <ModelPicker
+        <AgentModelPicker
           models={state.models}
           selectedModel={view.modelId}
           onSelect={(modelId) => handles.selectPaneModel(view.paneId, modelId)}
@@ -320,182 +318,4 @@ function renderWorkspacePane({
       onRegisterHandle={(handle) => handles.registerPaneHandle(view.paneId, handle)}
     />
   );
-}
-
-function ModelPicker({
-  models,
-  selectedModel,
-  onSelect,
-  loading,
-}: {
-  models: AgentModel[];
-  selectedModel: string;
-  onSelect: (id: string) => void;
-  loading: boolean;
-}) {
-  const [open, setOpen] = useState(false);
-  const controllerLabel = useActiveControllerLabel();
-  const active = models.find((model) => model.id === selectedModel) || null;
-  // Trigger shows the bare model id (rawId); the " · <controller>" suffix baked
-  // into `name` for multi-controller setups is dropped here.
-  const fallbackLabel = selectedModel || (models.length === 0 ? "No models" : "model");
-  const triggerLabel = loading
-    ? active?.rawId || active?.name || fallbackLabel || "Loading…"
-    : active?.rawId || active?.name || fallbackLabel;
-  const disabled = loading || models.length === 0;
-
-  // Group models by hardware controller so the dropdown shows one tab per
-  // controller and never repeats the controller id on every row.
-  const groups = useMemo(() => groupModelsByController(models), [models]);
-  const [activeControllerKey, setActiveControllerKey] = useState<string | null>(null);
-  const selectedKey = active ? controllerGroupKey(active) : null;
-  const currentKey = activeControllerKey ?? selectedKey ?? groups[0]?.key ?? null;
-  const currentGroup = groups.find((group) => group.key === currentKey) ?? groups[0] ?? null;
-
-  return (
-    <div
-      className="relative shrink-0"
-      onBlur={(event) => {
-        const nextTarget = event.relatedTarget;
-        if (nextTarget instanceof Node && event.currentTarget.contains(nextTarget)) return;
-        setOpen(false);
-      }}
-      onPointerDown={(event) => event.stopPropagation()}
-      onMouseDown={(event) => event.stopPropagation()}
-    >
-      <button
-        type="button"
-        onPointerDown={(event) => event.stopPropagation()}
-        onMouseDown={(event) => event.stopPropagation()}
-        onClick={() => {
-          if (disabled) return;
-          setOpen((value) => !value);
-        }}
-        disabled={disabled}
-        className="inline-flex !h-auto !min-h-0 !min-w-0 max-w-[160px] items-center gap-1 rounded-sm bg-transparent px-1 py-0.5 font-mono text-[length:var(--fs-xs)] text-(--dim) hover:text-(--fg) disabled:opacity-60"
-        title={active?.name || triggerLabel}
-      >
-        <span className="min-w-0 max-w-[132px] truncate">{triggerLabel}</span>
-        <ChevronDownIcon className="h-2.5 w-2.5 shrink-0" />
-      </button>
-      {open ? (
-        <div
-          className="absolute bottom-full right-0 z-[80] mb-1 w-80 overflow-hidden rounded-md border border-(--border) bg-[#151515] shadow-[0_12px_36px_rgba(0,0,0,0.65)]"
-          onPointerDown={(event) => event.stopPropagation()}
-          onMouseDown={(event) => event.stopPropagation()}
-        >
-          {groups.length > 1 ? (
-            <div className="flex items-center gap-1 overflow-x-auto border-b border-(--border) p-1.5">
-              {groups.map((group) => (
-                <button
-                  key={group.key}
-                  type="button"
-                  onClick={() => setActiveControllerKey(group.key)}
-                  className={`shrink-0 rounded px-2 py-1 font-mono text-[length:var(--fs-xs)] ${
-                    group.key === currentKey
-                      ? "bg-(--hover) text-(--fg)"
-                      : "text-(--dim) hover:text-(--fg)"
-                  }`}
-                >
-                  {group.name || controllerLabel || "local"}
-                </button>
-              ))}
-            </div>
-          ) : null}
-          <div className="max-h-72 overflow-y-auto p-1.5">
-            {(currentGroup?.models ?? []).map((model) => {
-              const isActive = model.id === selectedModel;
-              return (
-                <button
-                  key={model.id}
-                  type="button"
-                  onClick={() => {
-                    onSelect(model.id);
-                    setOpen(false);
-                  }}
-                  className={`flex w-full min-w-0 items-center gap-2 rounded px-2 py-2 text-left hover:bg-(--hover) ${
-                    isActive ? "bg-(--hover)" : ""
-                  }`}
-                >
-                  <span
-                    className={`h-1.5 w-1.5 shrink-0 rounded-full ${
-                      isActive ? "bg-(--accent)" : "bg-(--dim)/35"
-                    }`}
-                  />
-                  <span className="min-w-0 flex-1">
-                    <span className="block truncate text-xs text-(--fg)">
-                      {model.rawId || model.name}
-                    </span>
-                    <span className="mt-0.5 block truncate font-mono text-[length:var(--fs-xs)] text-(--dim)">
-                      {formatCompactNumber(model.contextWindow)} context
-                      {model.reasoning ? " · reasoning" : ""}
-                    </span>
-                  </span>
-                </button>
-              );
-            })}
-          </div>
-        </div>
-      ) : null}
-    </div>
-  );
-}
-
-type ModelGroup = { key: string; name: string; models: AgentModel[] };
-
-function controllerGroupKey(model: AgentModel): string {
-  return model.controllerUrl ?? model.controllerName ?? "primary";
-}
-
-function groupModelsByController(models: AgentModel[]): ModelGroup[] {
-  const groups = new Map<string, ModelGroup>();
-  for (const model of models) {
-    const key = controllerGroupKey(model);
-    const existing = groups.get(key);
-    if (existing) {
-      existing.models.push(model);
-    } else {
-      groups.set(key, { key, name: model.controllerName ?? "local", models: [model] });
-    }
-  }
-  return [...groups.values()];
-}
-
-function formatCompactNumber(value: number): string {
-  if (!Number.isFinite(value) || value <= 0) return "unknown";
-  if (value >= 1_000_000) return `${(value / 1_000_000).toFixed(value >= 10_000_000 ? 0 : 1)}M`;
-  if (value >= 1_000) return `${(value / 1_000).toFixed(value >= 10_000 ? 0 : 1)}K`;
-  return String(value);
-}
-
-function subscribeToControllerStorage(callback: () => void): () => void {
-  if (typeof window === "undefined") return () => undefined;
-  window.addEventListener("storage", callback);
-  return () => window.removeEventListener("storage", callback);
-}
-
-function computeActiveControllerLabel(): string | null {
-  const url = getStoredBackendUrl();
-  if (!url) return null;
-  const saved = loadSavedControllers();
-  if (saved.length === 0) return null;
-  const match = saved.find((entry) => entry.url === url);
-  return match?.name?.trim() || shortHost(url);
-}
-
-function useActiveControllerLabel(): string | null {
-  return useSyncExternalStore(
-    subscribeToControllerStorage,
-    computeActiveControllerLabel,
-    () => null,
-  );
-}
-
-function shortHost(url: string): string {
-  try {
-    const parsed = new URL(url);
-    return parsed.host || url;
-  } catch {
-    return url;
-  }
 }
